pyproject.toml
[tool.poetry]
name = "finmem-india"
version = "0.1.0"
description = "FinMem adaptation for Indian stock market (Nifty 50)"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
packages = [
    { include = "puppy" }
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.0"
numpy = "^1.26.0"
yfinance = "^0.2.36"
nsepy = "^0.8"
google-generativeai = "^0.8.0"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
toml = "^0.10.2"
click = "^8.1.7"
loguru = "^0.7.2"
transformers = "^4.37.0"
scikit-learn = "^1.4.0"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
aiohttp = "^3.9.0"
html5lib = "^1.1"
lxml = "^5.1.0"

[tool.poetry.group.dev.dependencies]
black = "^24.1.0"
isort = "^5.13.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 